requirements.txt
pylint			== 2.11.1
pytest			== 6.2.5
pytest-html		== 3.1.1
flask			== 2.0.2
flask-sqlalchemy	== 2.5.1
flask-wtf		== 0.15.1
WTForms			== 2.3.3
email-validator		== 1.1.3
flask_bcrypt		== 0.7.1
flask_login		== 0.5.0